requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
#awscli
flake8
#python-dotenv>=0.5.1
python-dotenv
## Sacred is a tool to help you configure, organize, log and reproduce experiments developed at IDSIA.
sacred
pymongo
artemis-ml

numpy
scikit-learn
pandas
scipy
